Fix the inversion handling problem
To clamp singular values of the deformation gradient to 0.577 if sigma(i) is less than 0.577, we had std::min(sigma(i), 0.577), but in reality, we wanted std::max(...) instead. We give the lower bound of 0.577 to the singular values to ensure that maximal stress is still applied on the tetrahedron under difficult configurations of St Venant Kirchhoff models where the stress would normally go to zero.

We also include a Neohookean constraint, which is not yet fully implemented. This type of constraint is still not usable. We should implement stable neohookean technique in future commits.